Se integra funcionalidad de creación de reglas de ajuste con validación en backend y diseño mejorado en frontend

+++ backend/src/main/java/com/codefactory/pricing_dinamico/domain/model/entities/Product.java
@@ -1,19 +1,22 @@
 package com.codefactory.pricing_dinamico.domain.model.entities;
 
+import javax.swing.*;
+import java.math.BigDecimal;
+
 public class Product {
     private Long id;
     private String sku;
     private String name;
     private String category;
-    private Long basePrice;
-    private Long minPrice;
-    private Long maxPrice;
+    private BigDecimal basePrice;
+    private BigDecimal minPrice;
+    private BigDecimal maxPrice;
     private ProductStatus productStatus;
 
     public Product() {
     }
 
-    public Product(Long id, String sku, String name, ProductStatus productStatus, Long basePrice, Long maxPrice, Long minPrice, String category) {
+    public Product(Long id, String sku, String name, ProductStatus productStatus, BigDecimal basePrice, BigDecimal maxPrice, BigDecimal minPrice, String category) {
         this.id = id;
         this.productStatus = productStatus;
         this.basePrice = basePrice;
@@ -50,11 +53,11 @@ public class Product {
         this.category = category;
     }
 
-    public Long getBasePrice() {
+    public BigDecimal getBasePrice() {
         return basePrice;
     }
 
-    public void setBasePrice(Long basePrice) {
+    public void setBasePrice(BigDecimal basePrice) {
         this.basePrice = basePrice;
     }
 
@@ -66,19 +69,19 @@ public class Product {
         this.productStatus = productStatus;
     }
 
-    public Long getMinPrice() {
+    public BigDecimal getMinPrice() {
         return minPrice;
     }
 
-    public void setMinPrice(Long minPrice) {
+    public void setMinPrice(BigDecimal minPrice) {
         this.minPrice = minPrice;
     }
 
-    public Long getMaxPrice() {
+    public BigDecimal getMaxPrice() {
         return maxPrice;
     }
 
-    public void setMaxPrice(Long maxPrice) {
+    public void setMaxPrice(BigDecimal maxPrice) {
         this.maxPrice = maxPrice;
     }
 }
